pyproject.toml
[tool.poetry]
name = "contacts-api-25"
version = "0.1.0"
description = ""
authors = ["Sasha"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.4"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.36"
psycopg2 = "^2.9.10"
python-multipart = "^0.0.17"
pydantic-settings = "^2.6.1"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
libgravatar = "^1.0.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
